Divide ScheduleTableRow component
